+++ packages/Dendrogram/src/package.ts
@@ -264,7 +264,7 @@ export class PackageFunctions {
     df: DG.DataFrame,
     colNameList: DG.ColumnList,
     @grok.decorators.param({type: 'string', options: {initialValue: 'euclidean', choices:['euclidean', 'manhattan']}}) distance: DistanceMetric = DistanceMetric.Euclidean,
-    @grok.decorators.param({options: {initialValue: 'ward', choices:['single', 'complete', 'average', 'weighted', 'centroid', 'median', 'ward']}})linkage: string,
+    @grok.decorators.param({options: {initialValue: 'ward', choices:['single', 'complete', 'average', 'weighted', 'median', 'ward']}})linkage: string,
   ): Promise<void> {
     const names = Array.isArray(colNameList) ? (colNameList as unknown as DG.Column[]).map((a) => a.name) : colNameList.names();
     await hierarchicalClusteringUI(df, names, distance, linkage);